Refactor Address class in Consultation package
