+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -10,10 +10,10 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.attendance.Attendance;
+import seedu.address.model.consultation.Address;
 import seedu.address.model.consultation.Day;
 import seedu.address.model.consultation.Time;
 import seedu.address.model.consultation.Type;
-import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.MatricNumber;
 import seedu.address.model.person.Name;
