+++ src/main/java/seedu/address/model/consultation/Consultation.java
@@ -4,7 +4,6 @@ import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 
 import java.util.Objects;
 
-import seedu.address.model.person.Address;
 import seedu.address.model.person.Name;
 
 /**
