+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -9,11 +9,11 @@ import java.util.stream.Collectors;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.Tasker;
 import seedu.address.model.attendance.Attendance;
+import seedu.address.model.consultation.Address;
 import seedu.address.model.consultation.Consultation;
 import seedu.address.model.consultation.Day;
 import seedu.address.model.consultation.Time;
 import seedu.address.model.consultation.Type;
-import seedu.address.model.person.Address;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.MatricNumber;
 import seedu.address.model.person.Name;
