+++ src/test/java/seedu/address/logic/commands/AddConsultTest.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.commands;
+
+public class AddConsultTest {
+}
