+++ src/test/java/seedu/address/logic/stubs/ModelStub.java
@@ -1,0 +1,4 @@
+package seedu.address.logic.stubs;
+
+public class ModelStub {
+}
