+++ src/test/java/seedu/address/model/consultation/AddressTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.consultation;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
