+++ src/test/java/seedu/address/testutil/ConsultationBuilder.java
@@ -1,10 +1,10 @@
 package seedu.address.testutil;
 
+import seedu.address.model.consultation.Address;
 import seedu.address.model.consultation.Consultation;
 import seedu.address.model.consultation.Day;
 import seedu.address.model.consultation.Time;
 import seedu.address.model.consultation.Type;
-import seedu.address.model.person.Address;
 import seedu.address.model.person.Name;
 
 /**
